CRUD CLIENT

+++ gestion-de-vente-facturation/src/main/java/com/emsi/gestion/de/vente/facturation/entities/Facture.java
@@ -16,7 +16,7 @@ public class Facture {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
-    private int Id;
+    private Long Id;
     private Date DateFacturation;
     private Double MontantTotal;
     private String StatutPaiement;
